+++ src/main/java/com/recuperai/domain/repository/ShopRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.recuperai.domain.entity.Shop;
 
-public interface ShopRepository extends JpaRepository<Shop, Long>{
+public interface ShopRepository extends JpaRepository<Shop, Integer>{
 
 }
